fix(notifications): delete trainee after 35 days of not completing

+++ src/api/Libraries/Modules/ECDLink.DataAccessLayer/Hierarchy/HierarchyEngine.cs
@@ -417,8 +417,10 @@ namespace ECDLink.DataAccessLayer.Hierarchy
                                .Where(x => x.UserId == userId.Value)
                                .OrderBy(x => x.Id)
                                .FirstOrDefault();
-
-            userHierarchyRepo.Delete(entity.Id);
+            if (entity != null)
+            {
+                userHierarchyRepo.Delete(entity.Id);
+            }
 
             return true;
         }
